+++ apps/fishing-map/features/modals/Modals.tsx
@@ -39,10 +39,9 @@ import EditWorkspaceModal from 'features/workspace/save/WorkspaceEditModal'
 import { selectIsWorkspaceReady } from 'features/workspace/workspace.selectors'
 import { setWorkspaceSuggestSave } from 'features/workspace/workspace.slice'
 import useSecretMenu, { useSecretKeyboardCombo } from 'hooks/secret-menu.hooks'
+import { getRouterRef } from 'router/router-ref'
 import { SAVE_WORKSPACE_BEFORE_LEAVE_KEY } from 'router/routes'
 import { ROUTE_PATHS } from 'router/routes.utils'
-
-import { getRouterRef } from '../../router'
 
 import styles from './Modals.module.css'
 
@@ -86,7 +85,7 @@ const TurningTidesMenuConfig = {
 const ResetWorkspaceConfig = {
   key: 'w',
   dispatchToggle: () => {
-    getRouterRef().navigate({ to: ROUTE_PATHS.HOME, search: {}, replace: true })
+    getRouterRef()?.navigate({ to: ROUTE_PATHS.HOME, search: {}, replace: true })
   },
 }
